+++ release_notes.md
@@ -2,7 +2,7 @@
 <!-- Please add your release notes in the following format:
 - My change description (#PR)
 -->
-- My change description (#PR)
+- Added analyzer suggesting best HttpClient usage practices. (#7718)
 
 **Release sprint:** Sprint 111
 [ [bugs](https://github.com/Azure/azure-functions-host/issues?q=is%3Aissue+milestone%3A%22Functions+Sprint+111%22+label%3Abug+is%3Aclosed) | [features](https://github.com/Azure/azure-functions-host/issues?q=is%3Aissue+milestone%3A%22Functions+Sprint+111%22+label%3Afeature+is%3Aclosed) ]